+++ backend/bookmarks/urls.py
@@ -5,6 +5,7 @@ from .views import (
     FolderListView,
     BookmarkListView,
     bookmark_create_from_telegram,
+    bookmark_create_from_web,
     BookmarkUpdateView,
 )
 
@@ -13,5 +14,6 @@ urlpatterns = [
     path("folders/", FolderListView.as_view()),
     path("bookmarks/", BookmarkListView.as_view()),
     path("bookmarks/create_from_telegram/", bookmark_create_from_telegram),
+    path("bookmarks/create/", bookmark_create_from_web),
     path("bookmarks/update/<int:pk>/", BookmarkUpdateView.as_view()),
 ]